Move ovo_energy base entity to separate module
